test: Members 관련 Rest Docs 완성
