+++ src/main/java/wooteco/subway/web/member/LoginMemberController.java
@@ -1,8 +1,13 @@
 package wooteco.subway.web.member;
 
+import java.net.URI;
+
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -11,6 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 import wooteco.subway.domain.member.Member;
 import wooteco.subway.service.member.MemberService;
 import wooteco.subway.service.member.dto.LoginRequest;
+import wooteco.subway.service.member.dto.MemberRequest;
 import wooteco.subway.service.member.dto.MemberResponse;
 import wooteco.subway.service.member.dto.TokenResponse;
 import wooteco.subway.service.member.dto.UpdateMemberRequest;
@@ -23,15 +29,23 @@ public class LoginMemberController {
 		this.memberService = memberService;
 	}
 
+	@GetMapping("/me")
+	public ResponseEntity<MemberResponse> getMemberOfMineBasic(@LoginMember Member member) {
+		return ResponseEntity.ok().body(MemberResponse.of(member));
+	}
+
+	@PostMapping("/me")
+	public ResponseEntity createMember(@RequestBody MemberRequest view) {
+		memberService.createMember(view.toMember());
+		return ResponseEntity
+			.noContent()
+			.build();
+	}
+
 	@PostMapping("/me/login")
 	public ResponseEntity<TokenResponse> login(@RequestBody LoginRequest param) {
 		String token = memberService.createToken(param);
 		return ResponseEntity.ok().body(new TokenResponse(token, "bearer"));
-	}
-
-	@GetMapping("/me")
-	public ResponseEntity<MemberResponse> getMemberOfMineBasic(@LoginMember Member member) {
-		return ResponseEntity.ok().body(MemberResponse.of(member));
 	}
 
 	@PutMapping("/me")
