+++ src/main/java/wooteco/subway/web/member/interceptor/BearerAuthInterceptor.java
@@ -27,7 +27,7 @@ public class BearerAuthInterceptor implements HandlerInterceptor {
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response,
         Object handler) {
         // TODO: Authorization 헤더를 통해 Bearer 값을 추출 (authExtractor.extract() 메서드 활용)
-        if (request.getMethod().equals("POST") && request.getRequestURI().equals("/members")) {
+        if (request.getMethod().equals("POST") && request.getRequestURI().equals("/me")) {
             return true;
         }
