+++ src/test/java/wooteco/subway/web/member/LoginMemberControllerTest.java
@@ -55,6 +55,7 @@ public class LoginMemberControllerTest {
         given(loginMemberMethodArgumentResolver.resolveArgument(any(), any(), any(),
             any())).willReturn(member);
         given(loginMemberMethodArgumentResolver.supportsParameter(any())).willReturn(true);
+        given(memberService.createToken(any())).willReturn("brownToken");
 
         this.mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext)
             .addFilter(new ShallowEtagHeaderFilter())
@@ -62,13 +63,30 @@ public class LoginMemberControllerTest {
             .build();
     }
 
+    @DisplayName("회원가입을 한다")
+    @Test
+    void createMember() throws Exception {
+        Member member = new Member(1L, TEST_USER_EMAIL, TEST_USER_NAME, TEST_USER_PASSWORD);
+        given(memberService.createMember(any())).willReturn(member);
+
+        String inputJson = "{\"email\":\"" + TEST_USER_EMAIL + "\"," +
+            "\"name\":\"" + TEST_USER_NAME + "\"," +
+            "\"password\":\"" + TEST_USER_PASSWORD + "\"}";
+
+        this.mockMvc.perform(RestDocumentationRequestBuilders.post("/me")
+            .content(inputJson)
+            .accept(MediaType.APPLICATION_JSON)
+            .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(status().isNoContent())
+            .andDo(print())
+            .andDo(LoginMemberDocumentation.createMember());
+    }
+
     @DisplayName("로그인을 시도하여 토큰을 얻는다.")
     @Test
     void tokenLogin() throws Exception {
         String inputJson = "{\"email\":\"" + TEST_USER_EMAIL + "\"," +
             "\"password\":\"" + TEST_USER_PASSWORD + "\"}";
-
-        given(memberService.createToken(any())).willReturn("brown");
 
         this.mockMvc.perform(RestDocumentationRequestBuilders.post("/me/login")
             .content(inputJson)
@@ -83,7 +101,7 @@ public class LoginMemberControllerTest {
     @Test
     void meBearer() throws Exception {
         this.mockMvc.perform(RestDocumentationRequestBuilders.get("/me")
-            .sessionAttr("loginMemberEmail", TEST_USER_EMAIL)
+            .header("Authorization", "Bearer brownToken")
             .accept(MediaType.APPLICATION_JSON)
             .contentType(MediaType.APPLICATION_JSON))
             .andDo(print())
@@ -97,7 +115,7 @@ public class LoginMemberControllerTest {
             "\"password\":\"" + "1234" + "\"}";
 
         this.mockMvc.perform(put("/me")
-            .sessionAttr("loginMemberEmail", TEST_USER_EMAIL)
+            .header("Authorization", "Bearer brownToken")
             .accept(MediaType.APPLICATION_JSON)
             .content(inputJson)
             .contentType(MediaType.APPLICATION_JSON))
@@ -109,8 +127,10 @@ public class LoginMemberControllerTest {
     @DisplayName("회원 정보 삭제")
     @Test
     void deleteMember() throws Exception {
+        given(memberService.createToken(any())).willReturn("brown");
+
         this.mockMvc.perform(RestDocumentationRequestBuilders.delete("/me")
-            .sessionAttr("loginMemberEmail", TEST_USER_EMAIL)
+            .header("Authorization", "Bearer brownToken")
             .accept(MediaType.APPLICATION_JSON)
             .contentType(MediaType.APPLICATION_JSON))
             .andDo(print())
